Do not let the substituted row count price a semi/anti join as expanding
`EnumeratorCheckerWithCosts::accept` resolves the join's kind and strictness before costing the
pair, but `computeJoinCost` ignored both and always used `estimateJoinedRows`, which returns
`max(lhs, rhs)` for a non-reliable equi join. Charging an unestimated relation
`unknown_relation_rows` therefore priced a semi/anti join over it as expanding to that surrogate,
although such a join only filters its preserved side and its recorded cardinality is clamped
accordingly.

With `query_plan_optimize_join_order_algorithm = 'dpsub'` and a conflict detector (the only
configuration that reorders semi/anti joins instead of swapping them), a selective `SEMI`/`ANTI`
filter over an unestimated relation was deferred behind the larger join. Measured on `k_a` (10
rows) semi-joined to an unestimated relation and inner-joined to `k_big` (1e6 rows, 100k per key):
the plan changed from filter-first to filter-last and peak RSS rose from 425-566 MB to 717-757 MB
over three runs each, because the 1e6-row intermediate is materialised before the filter. Results
are unchanged.

The cost increment is now capped at the preserved side's row count when a side's row count was
substituted, which is an unconditional upper bound on what a semi/anti join emits. Two narrower
variants were measured and rejected: taking the increment from `estimateJoinCardinality` prices
every anti join at 1 row without column statistics (1 instead of 1000 on a disjoint-key fixture),
and capping unconditionally also reorders fully estimated graphs, where it moves a non-selective
anti join ahead of a selective inner join. Costing a relation the optimizer cannot estimate is
what this change is about, so graphs whose relations are all estimated keep the plans and costs
they have today.

Only DPsub under CD-A/CD-C is affected: `evaluateJoin` always passes `JoinStrictness::All`, and
the greedy solver never receives a semi/anti graph because `conflictDetectorReordersSemiAnti`
requires DPsub as the sole algorithm. `rowsForJoinCost` returns `UInt64` so a row count never
round-trips through `double` (`estimated_rows` can be `UInt64::max()`, whose round-trip is
undefined).

Reported by the automated reviewer on the pull request.

+++ src/Processors/QueryPlan/Optimizations/joinOrderCommon.h
@@ -210,9 +210,9 @@ inline std::optional<UInt64> estimateJoinCardinality(
     return estimateJoinCardinality(left->estimated_rows, right->estimated_rows, selectivity, join_kind);
 }
 
-inline double rowsForJoinCost(const std::optional<UInt64> & estimated_rows, UInt64 unknown_relation_rows)
+inline UInt64 rowsForJoinCost(const std::optional<UInt64> & estimated_rows, UInt64 unknown_relation_rows)
 {
-    return static_cast<double>(estimated_rows.value_or(unknown_relation_rows));
+    return estimated_rows.value_or(unknown_relation_rows);
 }
 
 inline double computeJoinCost(
@@ -221,8 +221,8 @@ inline double computeJoinCost(
     const SelectivityEstimate & selectivity,
     UInt64 unknown_relation_rows)
 {
-    double lhs = rowsForJoinCost(left->estimated_rows, unknown_relation_rows);
-    double rhs = rowsForJoinCost(right->estimated_rows, unknown_relation_rows);
+    double lhs = static_cast<double>(rowsForJoinCost(left->estimated_rows, unknown_relation_rows));
+    double rhs = static_cast<double>(rowsForJoinCost(right->estimated_rows, unknown_relation_rows));
     return left->cost + right->cost + estimateJoinedRows(selectivity, lhs, rhs);
 }
